Added UI errors to resp
